split files, added summaries
